requirements.txt
# Web & core
flask==2.3.2
requests
python-dotenv

# Diffusion & model
diffusers
transformers
torch
safetensors
accelerate
peft

# Image I/O
Pillow

# Metrics & benchmarking
torchvision
torchmetrics[image]
torch-fidelity
lpips
tqdm
pandas
